Re-organize tools dir
