+++ Makefile
@@ -3,7 +3,7 @@ PY?=python3
 BASEDIR=$(CURDIR)
 TESTSDIR=$(BASEDIR)/.tests
 SCHEMASDIR=$(BASEDIR)/.schemas
-PYTHONPATH=$(TESTSDIR)
+export PYTHONPATH := $(BASEDIR):$(TESTSDIR)
 
 VERBOSE ?= 0
