+++ tools/constants.py
@@ -1,0 +1,8 @@
+
+JSON_FORMAT_KWARGS = {
+    'indent': 2,
+    'separators': (', ', ': '),
+    'sort_keys': True,
+}
+
+
